vigil-system/critic_agent/requirements.txt
google-adk
fastapi
uvicorn[standard]
a2a-sdk[http-server]
